Parse

Edit

SLiDE.edit_with — Method

edit_with(y::Dict{Any,Any}; kwargs...)
edit_with(df::DataFrame, editor::T) where T<:Edit
edit_with(df::DataFrame, lst::Array{T}) where T<:Edit
edit_with(df::DataFrame, x::Describe, file::T) where T<:File
edit_with(file::T, y::Dict{Any,Any}; kwargs...)
edit_with(files::Array{T,N} where N, y::Dict{Any,Any}; kwargs...) where T<:File

This function edits the input DataFrame df and returns the resultant DataFrame.

Arguments

df::DataFrame on which to perform the edit.

editor::T where T<:Edit: DataType containing information about which edit to perform. The following edit options are available and detailed below. If given a dictionary of edits, they will be made in this order:

SLiDE.Drop: Remove information from the DataFrame – either an entire column or rows containing specified values.

SLiDE.Rename: Change column name from -> to.

SLiDE.Group: Use to edit files containing data in successive dataframes with an identifying header cell or row.

SLiDE.Match: Extract values from the specified column into a column or columns based on the specified regular expression.

SLiDE.Melt: Normalize the dataframe by 'melting' columns into rows, lengthening the dataframe by duplicating values in the column on into new rows and defining 2 new columns:

var with header names from the original dataframe.

val with column values from the original dataframe.

SLiDE.Add: Add new column col filled with val.

SLiDE.Map: Define an output column containing values based on those in an input column. The mapping columns from -> to are contained in a .csv file in the coremaps directory. The columns input and from should contain the same values, as should output and to.

SLiDE.Replace: Replace values in col from -> to.

SLiDE.Operate: Perform an arithmetic operation across multiple DataFrame columns or rows.

SLiDE.Describe: This DataType is required when multiple DataFrames will be appended into one output file (say, if multiple sheets from an XLSX file are included). Before the DataFrames are appended, a column col will be added and filled with the value in the file descriptor.

SLiDE.Order: Rearranges columns in the order specified by cols and sets them to the specified type.

file::T where T <: File: Data file containing information to read.

files::Array{T} where T <: File: List of data files.

y::Dict{Any,Any}: Dictionary containing all editing structures among other values read from the yaml file. Dictionary keys must correspond EXACTLY with SLiDE.Edit DataType names, or the edits will not be made.

Keywords

shorten::Bool = false or shorten::Int: if an integer length is specified, the DataFrame will be shortened to the input value. This is meant to aid troubleshooting during development.

Returns

df::DataFrame including edit(s)

SLiDE.extrapolate_region — Function

extrapolate_region(df::DataFrame, yr::Array{Int64,1}; kwargs...)
extrapolate_year(df::DataFrame, set::Dict; kwargs...)

Arguments:

df::DataFrame that might be in need of extrapolation.

r::Pair = "md" => "dc": Pair indicating a region (r.first) to extrapolate to another region (r.second). A suggested regional extrapolation: MD data will be used to approximate DC data in the event that it is missing.

Keyword Argument:

overwrite::Bool = false: If data in the target region r.second is already present, should it be overwritten?

Returns:

df::DataFrame extrapolated in time.

SLiDE.extrapolate_year — Method

extrapolate_year(df::DataFrame, yr::Array{Int64,1}; kwargs...)
extrapolate_year(df::DataFrame, set::Dict; kwargs...)

Arguments:

df::DataFrame that might be in need of extrapolation.

yr::Array{Int64,1}: List of years overwhich extrapolation is possible (depending on the kwargs)

set::Dict containing list of years, identified by the key :yr.

Keyword Arguments:

backward::Bool = true: Do we extrapolate backward in time?

forward::Bool = true: Do we extrapolate forward in time?

Returns:

df::DataFrame extrapolated in time.

SLiDE.fill_zero — Method

fill_zero(keys_unique::NamedTuple; value_colnames)
fill_zero(keys_unique::NamedTuple, df::DataFrame)
fill_zero(df::DataFrame...)
fill_zero(d::Dict...)
fill_zero(keys_unique, d::Dict)

Arguments

keys_unique::Tuple: A list of arrays whose permutations should be included in the resultant dictionary.

keys_unique::NamedTuple: A list of arrays whose permutations should be included in the resultant dictionary. The NamedTuple's keys correspond to the DataFrame columns where they will be stored.

d::Dict...: The dictionary/ies to edit.

df::DataFrame...: The DataFrame(s) to edit.

Keyword Arguments

value_colnames::Any = :value: "value" column labels to add and set to zero when creating a new DataFrame. Default is :value.

Usage

This function can be used to fill zeros in either a dictionary or DataFrame.

Options for DataFrame editing:

If only (a) dictionary/ies is/are input, the dictionaries will be edited such that they all contain all permutations of their key values. All dictionaries in a resultant list of dictionaries will be the same length.

If a dictionary is input with a list of keys, it will be edited to ensure that it includes all permutations.

If only a list of keys is input, a new dictionary will be created, containing all key permutations with values initialized to zero.

Options for DataFrame editing:

If only (a) DataFrame(s) is/are input, the DataFrame(s) will be edited such that they all contain all permutations of their key values. All DataFrames in a resultant list of DataFrames will be the same length.

If a DataFrame is input with a NamedTuple, it will be edited to ensure that it includes all permutations of the NamedTuple's values.

If only a NamedTuple is input, a new DataFrame will be created, containing all key permutations with values initialized to zero.

Returns

d::Dict... if input included dictionaries and/or Tuples

df::DataFrame... if input included DataFrames and/or NamedTuples

SLiDE.filter_with — Method

filter_with(df::DataFrame, set::Any; kwargs...)

Arguments:

df::DataFrame to filter.

set::Dict or set::NamedTuple: Values to keep in the DataFrame.

Keyword Arguments:

extrapolate::Bool = false: Add missing regions/years to the DataFrame? If extrapolate is set to true, the following kwargs become relevant:

When extrapolating over years,

backward::Bool = true: Do we extrapolate backward in time?

forward::Bool = true: Do we extrapolate forward in time?

When extrapolating across regions,

r::Pair = "md" => "dc": Pair indicating a region (r.first) to extrapolate to another region (r.second). A suggested regional extrapolation: MD data will be used to approximate DC data in the event that it is missing.

overwrite::Bool = false: If data in the target region r.second is already present, should it be overwritten?

Returns:

df::DataFrame with only the desired keys.

Load

SLiDE.gams_to_dataframe — Method

gams_to_dataframe(xf::Array{String,1}; colpropertynames = false)

This function converts a GAMS map or set to a DataFrame, expanding sets into multiple rows.

Arguments

xf::Array{String,1}: A list of rows of text from a .map or a .set input file.

Keywords

colpropertynames = false: A user has the option to specify the column propertynames of the output DataFrame. If none are specified, a default of [missing, missing_1, ...] will be used, consistent with the default column headers for CSV.read() if column propertynames are missing.

Returns

df::DataFrame: A dataframe representation of the GAMS map or set.

SLiDE.load_from — Method

load_from(::Type{T}, df::DataFrame) where T <: Any

Load a DataFrame df into a structure of type T.

Note

This requires that all structure fieldnames are also DataFrame column names. Extra dataframe columns are acceptable, although that information will not be used.

Arguments

::Type{T} where T <: Any: Any DataType.

df::DataFrame: The DataFrame storing the information to store as a DataType.

Returns

x<:Any: The DataType specified as an argument.

lst::Array{T} where T<:Any: A list of elements of the DataType specified as an argument given a multi-row DataFrame.

Example

df = DataFrame(from = ["State"], to = ["region"])
load_from(Rename, df)

SLiDE.read_file — Method

read_file(file::String)
read_file(path::Array{String,1}, file<:File; kwargs...)
read_file(path::Array{String,1}, file::XLSXInput)
read_file(path::String, x::T) where T<:File
read_from(editor::T) where T<:Edit

This method to reads .csv mapping files required for editing. These files must be stored in the data/coremaps directory. It returns a .csv file.

Arguments

path::Array{String,1} or path::String: Path to file location; does not include file name.

file::String: Full path to file, including file name.

file<:File: A SLiDE DataType used to store information about a file. Options include:

SLiDE.CSVInput

SLiDE.XLSXInput

editor<:Edit: A SLiDE DataType used to store information about an edit to make in a DataFrame. Specifically, this function might be called for edit types that include the field file in reference to

SLiDE.Group

SLiDE.Map

Keywords

shorten::Bool = false or shorten::Int: if an integer length is specified, the DataFrame will be shortened to the input value. This is meant to aid troubleshooting during development.

Returns

df::DataFrame: If the input is a csv or xlsx file, this method will return a DataFrame.

yml::Dict{Any,Any}: If the input file is a yaml file, this method will return a dictionary. All keys that correspond with SLiDE DataStream DataTypes will be converted to (lists of) those types.

SLiDE.run_yaml — Method

run_yaml(filename::String)
run_yaml(filenames::Array{String,1})

This file runs (a) yaml file(s) if it includes the line Editable: true. If given a list of input files, the function will print a list of yaml files that were not marked as editable. A file might not be "editable" if SLiDE functionality cannot make all of the specified edits.

Arguments

filename::String or filenames::Array{String,1}: yaml file name (or list of propertynames) to run

Return

filename::String or filenames::Array{String,1}: yaml file name (or list of propertynames) that was/were not ran by the function because they were annotated with Editable: false

SLiDE.write_yaml — Method

write_yaml(path, file::XLSXInput)

This function reads an XLSX file and writes a new yaml file containing the information in each spreadsheet column. Sheet propertynames in the XLSX file correspond to the directory where new files will be printed (path/file.sheet/). Yaml files will be named after the text in the column's first row.

Arguments

path::String or path::Array{String,1}: XLSX file location. New yaml files will be printed here, as well.

file::XLSXInput or files::Array{XLSXInput,1}: XLSX file information (file name, sheet name, sheet range) or list of multiple sheets.

Returns

filenames::Array{String,1}: List of yaml files

Verify

SLiDE.benchmark! — Method

benchmark!(d_summ::Dict, k::Symbol, d_bench::Dict, d_calc::Dict;

SLiDE.compare_keys — Method

compare_keys(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1})

Arguments

df_lst::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

inds::Array{Symbol,1}: List of indicators that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and indicators.

SLiDE.compare_summary — Method

compare_summary(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1}; kwargs...)

Arguments

df_lst::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

inds::Array{Symbol,1}: List of indicators that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and indicators.

Keyword Argument

tol::Float64 = 1E-6: Tolerance used when determining whether values are equal. Default values is 1E-6.

SLiDE.compare_values — Method

compare_values(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1})

Arguments

df_lst::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

inds::Array{Symbol,1}: List of indicators that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and indicators.

Keyword Argument

tol::Float64 = 1E-6: Tolerance used when determining whether values are equal. Default values is 1E-6.

SLiDE.verify_over — Method

verify_over(df::DataFrame, col::Any; tol = 1E-6)