Parse

Edit

SLiDE._expand_range — Method

_expand_range()

SLiDE._filter_datastream — Method

Allows for basic filtering over years. Will need to expand to include regions.

SLiDE._find_constant — Method

SLiDE._group_by — Method

SLiDE._intersect_with — Method

SLiDE._print_status — Method

This function prints an editing status message.

SLiDE._sort_datastream — Method

_sort_datastream(df::DataFrame)

Returns the edited DataFrame, stored in a nicely-sorted order. This is most helpful for mapping and developing. Sorting isn't necessary and we could remove this function to save some time for users.

Load

SLiDE._delete_temporary — Method

Delete temporary files. These are pre-pended "_".

SLiDE._expand_set — Method

SLiDE._find_header — Method

SLiDE._load_as_type — Method

_load_as_type(::Type{Any}, entry, type::DataType)

Converts an entry to the required DataType

SLiDE._load_axis — Method

_load_axis(entry::Any)

SLiDE._load_case — Method

_load_case(entry::AbstractString)

Standardizes string identifiers that indicate a case change (upper-to-lower or vice-versa) for easier editing.

SLiDE._load_index — Method

SLiDE._load_path — Method

_load_path(d::Dict)

Edits directories containing a list of directories ending in a file name as one path.

SLiDE._read_csv — Method

SLiDE._remove_empty — Method

Ensure there are no empty columns (those containing all missing values)

SLiDE._remove_footer — Method

This function removes footer rows. These are identified by rows at the bottom of the sheet that are empty with the exception of the first column.

SLiDE._remove_header — Method

SLiDE._save_datastream — Method

SLiDE.split_gams — Method

Verify