CGE

SLiDE.Dataset — Type

mutable struct Dataset <: CGE
    name::String
    build::String
    step::String
    sector_level::Symbol
    eem::Bool
    save_build::Bool
    overwrite::Bool
end

Arguments

name::String: dataset identifier

build::String: Current step of the buildstream process: io or eem

step::String: Current substep of the buildstream. If build=io, these steps are partition, calibrate, share, disaggregate. If build=eem, these steps are ...

sector_level::Symbol: Aggregation level to use when reading BEA supply/use data.

eem::Bool: Flag indicating whether to include the Energy-Environment module. If eem=true, first build io supply/use data. Then build eem data.

save_build::Bool: Flag indicating decides whether to save the information at each build step. Setting save_build=true will add directories in the locations returned by SLiDE.datapath. This feature is particularly helpful for buildstream debugging.

overwrite::Bool: If data exists, do not read it. Build the data from scratch.

SLiDE.Parameter — Type

mutable struct Parameter <: CGE
    parameter::Symbol
    name::String
    index::Array{Symbol,1}
    units::String
end

Information about CGE parameters used in the model.

Arguments

parameter::Symbol: parameter variable abbreviation

name::String: parameter description

index::Array{Symbol,1}: Sets on which the parameter depends

units::String: parameter units