Regional Sharing

Divide data into regional component. This will guide how to break the national data into regional components by state or county.

SLiDE.share_region — Function

share_region(d::Dict, set::Dict; save_build = true, overwrite = false)

This function partitions BEA and Census Bureau data to use when disaggregating parameters from the national- to regional-level.

Arguments

dataset::Dataset identifier

d::Dict of model parameters

set::Dict of Arrays describing parameter indices (years, regions, goods, sectors, etc.)

Returns

d::Dict of model parameters

set::Dict of Arrays describing parameter indices (years, regions, goods, sectors, etc.)

CFS

SLiDE.share_rpc! — Function

rpc(r,g): Regional purchase coefficient

\[\rho_{r,g}^{cfs} = \begin{cases} \dfrac{\bar{d}_{r,g}} {\bar{d}_{r,g} - \bar{mn}_{r,g}} & r\neq uti, \bar{d}_{r,g} \neq \bar{mn}_{r,g} \\ 0.0 & r\neq uti, \bar{d}_{r,g} = \bar{mn}_{r,g} \\ 0.9 & r = uti \end{cases}\]

SLiDE._share_mrt0! — Function

mrt0(orig,dest,g): Interstate trade (CFS)

\[\bar{mrt}_{orig,dest,ng\ni g} = \left\{{cfs}\left(orig,dest,g\right) \;\vert\; orig\neq dest, \, g \right\}\]

SLiDE._share_d0! — Function

d0(r,g): Local supply-demand (CFS), trade that remains within the same region.

\[\bar{d}_{r,ng\ni g} = \left\{{cfs}\left(orig,dest,g\right) \;\vert\; orig=dest, \, g \right\}\]

Calling SLiDE._avg_ng returns $\bar{d}_{r,ng\in g}$.

SLiDE._share_xn0! — Function

xn0(r,g): National exports (CFS)

\[\bar{xn}_{r,ng\ni g} = \sum_{dest} \bar{mrt}_{orig,dest,ng\ni g}\]

Calling SLiDE._avg_ng returns $\bar{xn}_{r,ng\in g}$.

SLiDE._share_mn0! — Function

mn0(r,g): National demand (CFS)

\[\bar{mn}_{r,ng\ni g} = \sum_{orig} \bar{mrt}_{orig,dest,ng\ni g}\]

Calling SLiDE._avg_ng returns $\bar{mn}_{r,ng\in g}$.

SLiDE._avg_ng — Function

\[\bar{x}_{r,ng\in g} = \dfrac{\sum_g \bar{x}_{r,g}} {\text{length}(ng)}\]

GSP

SLiDE.share_region! — Function

region(yr,r,s): Regional share of value added

\[\alpha_{yr,r,s}^{gsp} = \begin{cases} \dfrac{ \bar{gdp}_{yr,r,s}} {\sum_{s'} \bar{gdp}_{yr,r,s'}} & \sum_{s'} \bar{gdp}_{yr,r,s'} \neq 0 \\ \dfrac{\sum_{s} \bar{gdp}_{yr,r ,s}} {\sum_{r',s} \bar{gdp}_{yr,r',s}} & \sum_{s'} \bar{gdp}_{yr,r,s'} = 0 \end{cases}\]

PCE

SLiDE.share_pce! — Function

pce(yr,r,g): Regional shares of final consumption

\[\alpha_{yr,r,g}^{pce} = \dfrac{\bar{pce}_{yr,r,g}}{\sum_{r'} \bar{pce}_{yr,r',g}}\]

SGF

SLiDE.share_sgf! — Function

sgf(yr,r,g): State Government Finance data.

\[\alpha_{yr,r,g}^{sgf} = \dfrac{\bar{sgf}_{yr,r,g}}{\sum_{r'} \bar{sgf}_{yr,r',g}}\]

Note: D.C. is not included in the original data set. Assume its SGF data is equal to Maryland's.

UTD

SLiDE.share_utd! — Function

utd(yr,r,g,t): Share of total trade by region.

\[\alpha_{yr,r,g,t}^{utd} = \begin{cases} \dfrac{ \bar{utd}_{yr,r ,g,t}} {\sum_{r'} \bar{utd}_{yr,r',g,t}} & notrd\ni g \\ \dfrac{\sum_{yr'} \bar{utd}_{yr',r ,g,t}} {\sum_{yr',r'} \bar{utd}_{yr',r',g,t}} & notrd\in g \end{cases}\]