Build

There are four steps required to build the cleaned input data for use in the model:

Partition

Calibrate

Share

Disaggregate

The buildstream process and all of the notation included in the SLiDE documentation here is meant to replicate the WiNDC buildstream. For more information, please reference:

Thomas F. Rutherford and Andrew Schreiber, "Tools for Open Source, Subnational CGE Modeling with an Illustrative Analysis of Carbon Leakage," J Global Econ Anal 4(2): 1-66.

SLiDE.build_data — Function

build_data(; kwargs...)

Keywords

save = true:

save = true (default) or save::String = "path/to/version": Save data in each build

step in the path returned by SLiDE.build_path.

save = false: the files at each step will not be saved.

overwrite = false: If data exists, do not read it. Rebuild the data from scratch.

Returns

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

Partition

Divide BEA supply/use data into parameters.

SLiDE.partition! — Function

partition!(d::Dict, set::Dict; kwargs...)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

save = true

overwrite = false

See SLiDE.build_data for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the

Calibrate

SLiDE.calibrate — Function

calibrate(d::Dict, set::Dict; save = true, overwrite = false)
calibrate(year::Int, d::Dict, set::Dict)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

year::Int: year for which to perform calibration

Keywords

save = true

overwrite = false

See SLiDE.build_data for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the calibration step.

Share

Divide data into regional component. This will guide how to break the national data into regional components by state or county.

SLiDE.share! — Function

share!(d::Dict, set::Dict; save = true, overwrite = false)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

save = true

overwrite = false

See SLiDE.build_data for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the sharing step.

Disaggregate

SLiDE.disagg! — Function

function disagg!(d::Dict, set::Dict; save = true, overwrite = false)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

save = true

overwrite = false

See SLiDE.build_data for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the disaggregation step