Sectoral Scaling

The build stream produces state-level (regionally) model parameters at the summary-level (sectorally). By default, there are 73 summary-level goods/sectors. These can be disaggregated into 409 detail-level goods/sectors using the blueNOTE sectoral scaling map.

Scaling features enable the following options:

Select a subset of summary-level or detail-level goods/sectors to examine.

Select a combination of summary- and detail-level goods/sectors.

Aggregate summary- and/or detail-level goods/sectors into those specified in a user-defined map.

Share

SLiDE.share_sector! — Function

share_sector!(dataset::String, d::Dict, set::Dict; kwargs...)

This function adds sharing information for the sectoral disaggregation to d[:sector].

It first reads and partitions (partition) detail-level BEA data to calculate gross output, $y_{yr,g}$, which is used to determine sectoral scaling for disaggregation. 409 detail-level codes are mapped to 73 summary-level codes using the blueNOTE sectoral scaling map, denoted here as $map_{gg\rightarrow g} = map_{ss\rightarrow s}$, where $gg$/$ss$ are summary-level goods/sectors and $g$/$s$ are detail-level goods/sectors.

A sectoral disaggregation parameter $\tilde{\delta}_{yr,gg \rightarrow g}$ can be defined:

\[\tilde{\delta}_{yr,gg \rightarrow g} = \dfrac {y_{yr,g} \circ map_{gg\rightarrow g}} {\sum_{gg} y_{yr,g} \circ map_{gg\rightarrow g}}\]

This process follows two major steps:

SLiDE.share_sector, which calculates $\tilde{\delta}_{yr,gg \rightarrow g}$ for all $(gg,g)$, (summary, detail) pairs.

SLiDE._combine_sector_levels, which selects relevant summary- and detail-level goods/sectors defined in set[:sector] and defines how to share the aggregate (summary)-level goods/sectors into this composite disaggregate-level set.

Arguments

dataset::String: dataset identifier

d::Dict of DataFrames containing regionally-disaggregated model parameters.

set::Dict of Arrays describing region, sector, final demand, etc. This should inlude a set set[:sector] listing the summary- and detail-level blueNOTE codes of interest to include in g.

Returns

d::Dict: regionally-disaggregated model parameters, updated to include d[:sector] = $\tilde{\delta}_{yr,gg \rightarrow g}$.

set::Dict of Arrays describing region, sector, final demand, etc.

SLiDE.share_sector — Function

share_sector(df::DataFrame)

This function calculates and returns a sectoral disaggregation parameter for all $(gg,g)$, (summary, detail) pairs.

\[\tilde{\delta}_{yr,gg \rightarrow g} = \dfrac {y_{yr,g} \circ map_{gg\rightarrow g}} {\sum_{gg} y_{yr,g} \circ map_{gg\rightarrow g}}\]

Example

Looking at a slice of the disaggregate-level $y_{yr,g}$ parameter,

julia> df = read_file(joinpath(SLIDE_DIR,"docs","src","assets","data","share_sector-y0_det.csv"))
6×3 DataFrame
│ Row │ yr    │ g       │ value   │
│     │ Int64 │ String  │ Float64 │
├─────┼───────┼─────────┼─────────┤
│ 1   │ 2007  │ ele_uti │ 292.105 │
│ 2   │ 2007  │ gas_uti │ 120.746 │
│ 3   │ 2007  │ wat_uti │ 10.331  │
│ 4   │ 2012  │ ele_uti │ 339.302 │
│ 5   │ 2012  │ gas_uti │ 78.708  │
│ 6   │ 2012  │ wat_uti │ 12.583  │

julia> df = SLiDE.share_sector(df)
6×4 DataFrame
│ Row │ yr     │ summary │ detail  │ value     │
│     │ Int64? │ String? │ String? │ Float64   │
├─────┼────────┼─────────┼─────────┼───────────┤
│ 1   │ 2007   │ uti     │ ele_uti │ 0.690259  │
│ 2   │ 2007   │ uti     │ gas_uti │ 0.285329  │
│ 3   │ 2007   │ uti     │ wat_uti │ 0.0244127 │
│ 4   │ 2012   │ uti     │ ele_uti │ 0.787988  │
│ 5   │ 2012   │ uti     │ gas_uti │ 0.18279   │
│ 6   │ 2012   │ uti     │ wat_uti │ 0.0292225 │

This ensures $\sum_{g}\tilde{\delta}_{yr,gg \rightarrow g} = 1$.

julia> combine_over(df, :detail)
2×3 DataFrame
│ Row │ yr     │ summary │ value   │
│     │ Int64? │ String? │ Float64 │
├─────┼────────┼─────────┼─────────┤
│ 1   │ 2007   │ uti     │ 1.0     │
│ 2   │ 2012   │ uti     │ 1.0     │

SLiDE._combine_sector_levels — Function

_combine_sector_levels(df::DataFrame, x::AbstractArray; kwargs...)

Given a sharing DataFrame that includes level-mapping and a list of all parameters that should be included at the disaggregate level, this function returns a complete mapping DataFrame to scale between mixed levels.

Arguments

df::DataFrame: sharing DataFrame with columns mapped between levels.

x::AbstractArray: composite of all levels to include on the disaggregate level.

Keywords

scheme::Pair = :summary=>:detail: columns in df to disaggregate from $\rightarrow$ to.

Returns

df::DataFrame: complete mapping DataFrame to scale between mixed levels. The (aggregate-level, composite) columns are named (aggr,disagg) such that the sum over the composite column is one.

Example

Focusing on a slice of $\tilde{\delta}_{yr,gg \rightarrow g}$,

julia> df = read_file(joinpath(SLIDE_DIR,"docs","src","assets","data","combine_sector_levels-sector_share.csv"))
8×4 DataFrame
│ Row │ yr    │ summary │ detail  │ value     │
│     │ Int64 │ String  │ String  │ Float64   │
├─────┼───────┼─────────┼─────────┼───────────┤
│ 1   │ 2012  │ oil     │ oil_oil │ 1.0       │
│ 2   │ 2012  │ pet     │ oth_pet │ 0.0299492 │
│ 3   │ 2012  │ pet     │ pav_pet │ 0.015931  │
│ 4   │ 2012  │ pet     │ ref_pet │ 0.941366  │
│ 5   │ 2012  │ pet     │ shn_pet │ 0.0127534 │
│ 6   │ 2012  │ uti     │ ele_uti │ 0.787988  │
│ 7   │ 2012  │ uti     │ gas_uti │ 0.18279   │
│ 8   │ 2012  │ uti     │ wat_uti │ 0.0292225 │

julia> x = ["oil","ref_pet","uti","ele_uti"];

julia> (df, x) = SLiDE._combine_sector_levels(df, x);

julia> df
5×4 DataFrame
│ Row │ yr    │ aggr   │ disagg  │ value     │
│     │ Int64 │ String │ String  │ Float64   │
├─────┼───────┼────────┼─────────┼───────────┤
│ 1   │ 2012  │ oil    │ oil     │ 1.0       │
│ 2   │ 2012  │ pet    │ pet     │ 0.0586336 │
│ 3   │ 2012  │ pet    │ ref_pet │ 0.941366  │
│ 4   │ 2012  │ uti    │ ele_uti │ 0.787988  │
│ 5   │ 2012  │ uti    │ uti     │ 0.212012  │

In the case that set[:sector]($g$) includes a disaggregate-level sector without its corresponding aggregate-level code (e.g. ref_pet without pet, as is the case in this example), the aggregate-level code will be added.

julia> x
5-element Array{String,1}:
 "oil"
 "pet"
 "ref_pet"
 "ele_uti"
 "uti"

This ensures that $\sum_{g}\tilde{\delta}_{yr,gg \rightarrow g} = 1$.

julia> combine_over(df, :disagg)
3×3 DataFrame
│ Row │ yr    │ aggr   │ value   │
│     │ Int64 │ String │ Float64 │
├─────┼───────┼────────┼─────────┤
│ 1   │ 2012  │ oil    │ 1.0     │
│ 2   │ 2012  │ pet    │ 1.0     │
│ 3   │ 2012  │ uti    │ 1.0     │

Disaggregate

SLiDE.disagg_sector! — Function

disagg_sector!(dataset::String, d::Dict, set::Dict; kwargs...)

This function performs the sectoral disaggregation for all of the model parameters.

Taxes ($ta$, $tm$, $ty$) are disaggregated using SLiDE._disagg_sector_map.

All other parameters are disaggregated using SLiDE._disagg_sector_share.

Since disaggregate-level data is available only for 2007 and 2012, the map d[:sector] must be extended using SLiDE.map_year, so that

\[\tilde{\delta}_{yr,gg \rightarrow g} = \begin{cases} \tilde{\delta}_{2007,gg \rightarrow g} & yr \leq 2009 \\ \tilde{\delta}_{2012,gg \rightarrow g} & yr > 2009 \end{cases}\]

where $gg$, $ss$ represent aggregate-level goods and sectors and $g$, $s$ represent disaggregate-level goods and sectors.

Arguments

dataset::String: dataset identifier

d::Dict of DataFrames containing the model parameters. This must include d[:sector], $\tilde{\delta}_{yr,gg \rightarrow g}$, of sectoral shares.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

scheme::Pair = :aggr=>:disagg: columns in dfmap to disaggregate from $\rightarrow$ to

Returns

d::Dict of DataFrames containing disaggregated model parameters.

set::Dict of Arrays describing region, sector, final demand, etc.

SLiDE._disagg_sector_share — Function

_disagg_sector_share(df::DataFrame, dfmap::DataFrame; kwargs...)

This function maps df from the aggregate to the disaggregate level and multiplies by the sectoral sharing defined in dfmap. For a parameter $\bar{z}$,

\[\begin{aligned} \bar{z}_{yr,r,g} = \sum_{gg} \left( \bar{z}_{yr,r,gg} \cdot \tilde{\delta}_{yr,gg \rightarrow g} \right) \end{aligned}\]

where $gg$, $ss$ represent aggregate-level goods and sectors and $g$, $s$ represent disaggregate-level goods and sectors.

For parameters, such as sectoral output, $ys_{yr,r,ss,gg}$, and intermediate demand, $id_{yr,r,gg,ss}$, that depend on both goods and sectors, SLiDE._compound_for is used to produce a compound disaggregation map.

\[\begin{aligned} \bar{id}_{yr,r,g,s} &= \sum_{gg,ss} \left( \bar{id}_{yr,r,gg,ss} \cdot \tilde{\delta}_{yr,gg \rightarrow g,ss \rightarrow s} \right) \\ \bar{ys}_{yr,r,s,g} &= \sum_{gg,ss} \left( \bar{ys}_{yr,r,ss,gg} \cdot \tilde{\delta}_{yr,ss \rightarrow s,gg \rightarrow g} \right) \end{aligned}\]

Arguments

df::DataFrame: parameter to share

dfmap::DataFrame: sharing

Keywords

scheme::Pair = :aggr=>:disagg: columns in dfmap to disaggregate from $\rightarrow$ to

key: If a value is given, print a message indicating that key is being shared.

Returns

df::DataFrame: shared parameter

SLiDE._compound_for — Function

_compound_for(df::DataFrame, col::AbstractArray; kwargs...)

This function produces a map to disaggregate parameters that include both goods and sectors:

\[\tilde{\delta}_{yr,gg \rightarrow g, ss \rightarrow s} = \tilde{\delta}_{yr,gg \rightarrow g} \cdot \tilde{\delta}_{yr, ss \rightarrow s}\]

where $gg$, $ss$ represent aggregate-level goods and sectors and $g$, $s$ represent disaggregate-level goods and sectors.

In the case that we are sharing across both goods and sectors in one data set, this function generates a dataframe with these sharing parameters through the following process:

Multiply shares for all ($gg\rightarrow g$,$ss\rightarrow s$) combinations.

Address the case of when aggregate-level goods and sectors are the same ($gg=ss$):

If $g = s$, sum all of the share values.

If $g\neq s$, drop these values.

Example

julia> df = read_file(joinpath(SLIDE_DIR,"docs","src","assets","data","compound_for-sector_share.csv"))
3×4 DataFrame
│ Row │ yr    │ aggr   │ disagg  │ value    │
│     │ Int64 │ String │ String  │ Float64  │
├─────┼───────┼────────┼─────────┼──────────┤
│ 1   │ 2012  │ oil    │ oil     │ 1.0      │
│ 2   │ 2012  │ uti    │ ele_uti │ 0.787988 │
│ 3   │ 2012  │ uti    │ uti     │ 0.212012 │

First, multiply shares for all ($gg\rightarrow g$,$ss\rightarrow s$) combinations:

9×6 DataFrame
│ Row │ yr     │ aggr_g  │ disagg_g │ aggr_s  │ disagg_s │ value     │
│     │ Int64? │ String? │ String?  │ String? │ String?  │ Float64   │
├─────┼────────┼─────────┼──────────┼─────────┼──────────┼───────────┤
│ 1   │ 2012   │ oil     │ oil      │ oil     │ oil      │ 1.0       │
│ 2   │ 2012   │ oil     │ oil      │ uti     │ ele_uti  │ 0.787988  │
│ 3   │ 2012   │ oil     │ oil      │ uti     │ uti      │ 0.212012  │
│ 4   │ 2012   │ uti     │ ele_uti  │ oil     │ oil      │ 0.787988  │
│ 5   │ 2012   │ uti     │ ele_uti  │ uti     │ ele_uti  │ 0.620925  │
│ 6   │ 2012   │ uti     │ ele_uti  │ uti     │ uti      │ 0.167063  │
│ 7   │ 2012   │ uti     │ uti      │ oil     │ oil      │ 0.212012  │
│ 8   │ 2012   │ uti     │ uti      │ uti     │ ele_uti  │ 0.167063  │
│ 9   │ 2012   │ uti     │ uti      │ uti     │ uti      │ 0.0449492 │

Now, for aggr_g==aggr_s, sum all share values for disagg_g==disagg_s and drop values if disagg_g!=disagg_s.

3×6 DataFrame
│ Row │ yr     │ aggr_g  │ disagg_g │ aggr_s  │ disagg_s │ value    │
│     │ Int64? │ String? │ String?  │ String? │ String?  │ Float64  │
├─────┼────────┼─────────┼──────────┼─────────┼──────────┼──────────┤
│ 1   │ 2012   │ oil     │ oil      │ oil     │ oil      │ 1.0      │
│ 2   │ 2012   │ uti     │ ele_uti  │ uti     │ ele_uti  │ 0.787988 │
│ 3   │ 2012   │ uti     │ uti      │ uti     │ uti      │ 0.212012 │

The resulting output is:

julia> SLiDE._compound_for(df, [:g,:s])
7×6 DataFrame
│ Row │ yr     │ aggr_g  │ disagg_g │ aggr_s  │ disagg_s │ value    │
│     │ Int64? │ String? │ String?  │ String? │ String?  │ Float64  │
├─────┼────────┼─────────┼──────────┼─────────┼──────────┼──────────┤
│ 1   │ 2012   │ oil     │ oil      │ oil     │ oil      │ 1.0      │
│ 2   │ 2012   │ oil     │ oil      │ uti     │ ele_uti  │ 0.787988 │
│ 3   │ 2012   │ oil     │ oil      │ uti     │ uti      │ 0.212012 │
│ 4   │ 2012   │ uti     │ ele_uti  │ oil     │ oil      │ 0.787988 │
│ 5   │ 2012   │ uti     │ uti      │ oil     │ oil      │ 0.212012 │
│ 6   │ 2012   │ uti     │ ele_uti  │ uti     │ ele_uti  │ 0.787988 │
│ 7   │ 2012   │ uti     │ uti      │ uti     │ uti      │ 0.212012 │

SLiDE._disagg_sector_map — Function

_disagg_sector_map(df::DataFrame, dfmap::DataFrame; kwargs...)

This function maps df from the aggregate to the disaggregate level such that all disaggregate-level goods/sectors have the same value as at the aggregate-level. For a parameter $\bar{z}$,

\[\bar{z}_{yr,r,g} = \bar{z}_{yr,r,gg} \circ map_{gg \rightarrow g}\]

where $gg$, $ss$ represent aggregate-level goods and sectors and $g$, $s$ represent disaggregate-level goods and sectors.

Arguments

df::DataFrame: parameter to map

dfmap::DataFrame: mapping $map_{gg \rightarrow g}$ or $\tilde{\delta}_{yr,gg \rightarrow g}$

Keywords

scheme::Pair = :aggr=>:disagg: columns in dfmap to disaggregate from $\rightarrow$ to

key: If a value is given, print a message indicating that key is being mapped.

Returns

df::DataFrame: mapped parameter

Aggregate

SLiDE.aggregate_sector! — Function

aggregate_sector!(dataset::String, d::Dict, set::Dict; kwargs...)

This function performs the sectoral aggregation for all of the model parameters.

Taxes ($ta$, $tm$, $ty$) are aggregated using SLiDE._aggregate_tax_with!.

All other parameters are disaggregated using SLiDE._aggregate_sector_map.

Arguments

dataset::String: dataset identifier

d::Dict of DataFrames containing the model parameters.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

scheme::Pair = :disagg=>:aggr: columns in dfmap to aggregate from $\rightarrow$ to

path::String: path to the defaul sectoral aggregation mapping file.

Returns

d::Dict of DataFrames containing disaggregated model parameters.

set::Dict of Arrays describing region, sector, final demand, etc.

SLiDE._aggregate_tax_with! — Function

_aggregate_tax_with!(d::Dict, dfmap, kt, k; kwargs...)

This function is applied to the following tax rates, which are scaled by a the corresponding parameters. It returns the aggregated tax rate as well as the aggregated scaling parameter.

ta(yr,r,g), absorption taxes, scaled by domestic absorption, a(yr,r,g):

\[\begin{aligned} \bar{a}_{yr,r,g} &= \bar{a}_{yr,r,gg} \circ map_{gg\rightarrow g} \\ \bar{ta}_{yr,r,g} &= \dfrac {\left(\bar{ta}_{yr,r,gg} \cdot \bar{a}_{yr,r,gg} \right) \circ map_{gg\rightarrow g}} { \bar{a}_{yr,r,g}} \end{aligned}\]

tm(yr,r,g), import taxes, scaled by foreign imports, m(yr,r,g):

\[\begin{aligned} \bar{m}_{yr,r,g} &= \bar{m}_{yr,r,gg} \circ map_{gg\rightarrow g} \\ \bar{tm}_{yr,r,g} &= \dfrac {\left(\bar{tm}_{yr,r,gg} \cdot \bar{m}_{yr,r,gg} \right) \circ map_{gg\rightarrow g}} { \bar{m}_{yr,r,g}} \end{aligned}\]

ty(yr,r,s), production taxes, scaled by total regional sectoral output, ys(yr,r,s,g):

\[\begin{aligned} \bar{ys}_{yr,r,s,g} &= \bar{ys}_{yr,r,ss,gg} \circ map_{ss\rightarrow s, gg\rightarrow g} \\ \bar{ty}_{yr,r,s} &= \dfrac {\left(\bar{ty}_{yr,r,ss} \cdot \sum_{gg} \bar{ys}_{yr,r,ss,gg} \right) \circ map_{ss\rightarrow s}} { \sum_{g} \bar{ys}_{yr,r,s ,g}} \end{aligned}\]

where $gg$, $ss$ represent disaggregate-level goods and sectors and $g$, $s$ represent aggregate-level goods and sectors.

Arguments

d::Dict of model parameters

dfmap::DataFrame: mapping $map_{gg\rightarrow g}$

kt::Symbol: tax parameter key

k::Symbol: scaling parameter key

Keywords

scheme::Pair = :disagg=>:aggr: columns in dfmap to aggregate from $\rightarrow$ to

Returns

d[kt]::DataFrame: mapped tax parameter

d[k]::DataFrame: mapped scaling parameter

SLiDE._aggregate_sector_map — Function

_aggregate_sector_map(df::DataFrame, dfmap::DataFrame; kwargs...)

This function aggregates a parameter in df according to the input map dfmap. For a parameter $\bar{z}$,

\[\bar{z}_{yr,r,g} = \sum_{gg} \left(\bar{z}_{yr,r,gg} \circ map_{gg\rightarrow g} \right)\]

where $gg$, $ss$ represent disaggregate-level goods and sectors and $g$, $s$ represent aggregate-level goods and sectors.

Keywords

scheme::Pair = :disagg=>:aggr: columns in dfmap to aggregate from $\rightarrow$ to

key: If a value is given, print a message indicating that key is being mapped.