Parse

Load

SLiDE.gams_to_dataframe — Method

gams_to_dataframe(xf::Array{String,1}; colnames = false)

This function converts a GAMS map or set to a DataFrame, expanding sets into multiple rows.

Arguments

xf::Array{String,1}: A list of rows of text from a .map or a .set input file.

Keywords

colnames = []: A user has the option to specify the column propertynames of the output DataFrame. If none are specified, a default of [missing, missing_1, ...] will be used, consistent with the default column headers for CSV.read() if column propertynames are missing.

Returns

df::DataFrame: A dataframe representation of the GAMS map or set.

SLiDE.load_from — Method

load_from(::Type{T}, df::DataFrame) where T <: Any

Load a DataFrame df into a structure of type T.

Note

This requires that all structure fieldnames are also DataFrame column names. Extra dataframe columns are acceptable, although that information will not be used.

Arguments

::Type{T} where T <: Any: Any DataType.

df::DataFrame: The DataFrame storing the information to store as a DataType.

Returns

x<:Any: The DataType specified as an argument.

lst::Array{T} where T<:Any: A list of elements of the DataType specified as an argument given a multi-row DataFrame.

Example

df = DataFrame(from = ["State"], to = ["region"])
load_from(Rename, df)

SLiDE.read_file — Method

read_file(file::String)
read_file(path::Array{String,1}, file<:File; kwargs...)
read_file(path::Array{String,1}, file::XLSXInput)
read_file(path::String, x::T) where T<:File
read_file(editor::T) where T<:Edit

This method to reads .csv mapping files required for editing. These files must be stored in the data/coremaps directory. It returns a .csv file.

Arguments

path::Array{String,1} or path::String: Path to file location; does not include file name.

file::String: Full path to file, including file name.

file<:File: A SLiDE DataType used to store information about a file. Options include:

SLiDE.CSVInput

SLiDE.XLSXInput

SLiDE.GAMSInput

SLiDE.SetInput

SLiDE.DataInput

editor<:Edit: A SLiDE DataType used to store information about an edit to make in a DataFrame. Specifically, this function might be called for edit types that include the field file in reference to

SLiDE.Group

SLiDE.Map

Returns

df::DataFrame: If the input is a csv or xlsx file, this method will return a DataFrame.

yml::Dict{Any,Any}: If the input file is a yaml file, this method will return a dictionary. All keys that correspond with SLiDE DataStream DataTypes will be converted to (lists of) those types.

SLiDE.run_yaml — Method

run_yaml(filename::String)
run_yaml(filenames::Array{String,1})

This file runs (a) yaml file(s) if it includes the line Editable: true. If given a list of input files, the function will print a list of yaml files that were not marked as editable. A file might not be "editable" if SLiDE functionality cannot make all of the specified edits.

Arguments

filename::String or filenames::Array{String,1}: yaml file name (or list of propertynames) to run

Return

filename::String or filenames::Array{String,1}: yaml file name (or list of propertynames) that was/were not ran by the function because they were annotated with Editable: false

SLiDE.write_yaml — Method

write_yaml(path, file::XLSXInput)

This function reads an XLSX file and writes a new yaml file containing the information in each spreadsheet column. Sheet propertynames in the XLSX file correspond to the directory where new files will be printed (path/file.sheet/). Yaml files will be named after the text in the column's first row.

Arguments

path::String or path::Array{String,1}: XLSX file location. New yaml files will be printed here, as well.

file::XLSXInput or files::Array{XLSXInput,1}: XLSX file information (file name, sheet name, sheet range) or list of multiple sheets.

Returns

filenames::Array{String,1}: List of yaml files

Verify

SLiDE.benchmark_against — Method

benchmark_against(df_calc, df_bench; kwargs...)

SLiDE.compare_keys — Method

compare_keys(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1})

Arguments

df_lst::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

inds::Array{Symbol,1}: List of IDs that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and IDs.

SLiDE.compare_summary — Method

compare_summary(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1}; kwargs...)

Arguments

df::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

d::Array{Dict,1}: Array of dictonaries of DataFrames to compare.

id::Array{Symbol,1}: List of IDs that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and IDs.

Keywords

tol::Float64 = 1E-6: Tolerance used when determining whether values are equal. Default values is 1E-6.

complete_summary::Bool = false: Should we include the full summary DataFrame or only the problematic rows (i.e., ones where either equal_keys = false or equal_values = false)?

Returns

df::DataFrame: Summary DataFrame including the index columns and values from the original df, marked to indicate which input DataFrame they came from, the reldiff between the values, and booleans equal_key (are values for the index present in both DataFrames?) and equal_value (are the values the same within the given tol?)

d::Dict: Dictionary of summary DataFrames

SLiDE.compare_values — Method

compare_values(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1})

Arguments

df_lst::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

inds::Array{Symbol,1}: List of IDs that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and IDs.

Keyword Argument

tol::Float64 = 1E-6: Tolerance used when determining whether values are equal. Default values is 1E-6.

SLiDE.verify_over — Method

verify_over(df::DataFrame, col::Any; tol = 1E-6)