Parse

Load

Verify

SLiDE.benchmark_against — Method

benchmark_against(df_calc, df_bench; kwargs...)

SLiDE.compare_keys — Method

compare_keys(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1})

Arguments

df_lst::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

inds::Array{Symbol,1}: List of IDs that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and IDs.

SLiDE.compare_summary — Method

compare_summary(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1}; kwargs...)

Arguments

df::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

d::Array{Dict,1}: Array of dictonaries of DataFrames to compare.

id::Array{Symbol,1}: List of IDs that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and IDs.

Keywords

tol::Float64 = 1E-6: Tolerance used when determining whether values are equal. Default values is 1E-6.

complete_summary::Bool = false: Should we include the full summary DataFrame or only the problematic rows (i.e., ones where either equal_keys = false or equal_values = false)?

Returns

df::DataFrame: Summary DataFrame including the index columns and values from the original df, marked to indicate which input DataFrame they came from, the reldiff between the values, and booleans equal_key (are values for the index present in both DataFrames?) and equal_value (are the values the same within the given tol?)

d::Dict: Dictionary of summary DataFrames

SLiDE.compare_values — Method

compare_values(df_lst::Array{DataFrame,1}, inds::Array{Symbol,1})

Arguments

df_lst::Array{DataFrame,1}: List of DataFrames to compare. These must all share the same column names.

inds::Array{Symbol,1}: List of IDs that describe each DataFrame and track which values/keys are present in each DataFrame. There must be an equal number of input DataFrames and IDs.

Keyword Argument

tol::Float64 = 1E-6: Tolerance used when determining whether values are equal. Default values is 1E-6.

SLiDE.verify_over — Method

verify_over(df::DataFrame, col::Any; tol = 1E-6)