<!– # Build

SLiDE._delete_empty! — Method

_delete_empty!(d::Dict)

This function removes dictionary entries with empty values.

SLiDE._edit_from_yaml — Method

SLiDE._inp_key — Method

_inp_key(x::Any)

This function is a standardized method for generating dictionary keys for SLiDE.read_from based on the type of information that is being read.

SLiDE._read_from_dir — Method

_read_from_dir(dir::String; kwargs...)

This function reads all of the files from the input directory and returns the contents of those of the specified extension.

Arguments

dir::String: Relative path to directory to read

Keywords

ext::String = ".csv: File extension to read and return

run_bash::Bool = false: If there's a shell script in dir, run it to generate/update directory contents before reading the files.

Returns

d::Dict{Symbol,Any} of file contents where the key references the source file name.

SLiDE._read_from_yaml — Method

_read_from_yaml(path::String)

SLiDE._run_bash — Method

_run_bash(dir::String, files)
_run_bash(path::String)

This function runs a shell script if one is found.

Arguments

path::String: Relative path to a specific shell script to run or to a directory that might contain a shell script.

file::String or files::Array{String,1}: A list of files in the specified directory.

Returns

files::Array{String,1}: An updated list of files in the specified directory after running the shell script if a list of files is given as an argument.

SLiDE.build_parameters — Method

build_parameters(subset::String)

Arguments

subset::String: Internally-passed parameter indicating the type of information to save (set, parameter, or build step)

Returns

d::Dict{Symbol,SLiDE.Parameter} of Parameters relevant to the specified data subset. The dictionary key is consistent the value's field parameter.

SLiDE.data_path — Method

data_path(dataset::String)

Arguments

dataset::String: Dataset identifier

Returns

dir::String = /path/to/dataset

SLIDE_DIR is the path to the location of the SLiDE.jl package on the user's machine.

The default dataset identifier is state_model. This dataset includes all U.S. states and summary-level sectors and goods.

SLiDE.filter_build! — Method

filter_build!(subset::String, d::Dict; kwargs...)

This function filters d to contain only keys relevant to the specified subset. This avoids cluttering a saved directory with superfluous parameters that may have been calculated at intermediate steps.

If filtering DataFrames, this reorders the DataFrame indices. This is important when importing parameters into JuMP models when calibrating or modeling.

Arguments

subset::String: Internally-passed parameter indicating the type of information to save (set, parameter, or build step)

d::Dict of DataFrames containing data for the specified data subset

Returns

d::Dict filtered to include only relevant parameters.

SLiDE.read_build — Method

read_build(dataset::String, subset::String; kwargs...)

This function reads data from the specified subset if this information has previously been generated and saved.

Arguments

dataset::String: Dataset identifier

subset::String: Internally-passed parameter indicating the type of information to save (set, parameter, or build step)

Keywords

overwrite::Bool = true: If the user would like to re-generate the subset of data in specified dataset, delete the directory. The information must now be re-calculated to repopulate the subset directory.

Returns

d::Dict{Symbol,DataFrame} if reading parameters or d::Dict{Any,Array} if reading sets

SLiDE.sub_path — Method

sub_path(dataset::String, subset::String; kwargs...)

Arguments

dataset::String: Dataset identifier

subset::String: Internally-passed parameter indicating the type of information to save (set, parameter, or build step)

Returns

path::String: Standardized location indicating where to save intermediate files. Here, /path/to/dataset is returned by SLiDE.data_path

If saving build steps: /path/to/dataset/build/<build_step>. For the For the four build steps, these are: partition, calibrate, share, and build.

Model input parameters: /path/to/dataset/parameters

Parameter indices: /path/to/dataset/sets

SLiDE.write_build! — Method

write_build(dataset::String, subset::String, d::Dict; kwargs...)

This function filters the contents of the input dictionary d to include only relevant files using SLiDE.filter_build! and writes set lists and parameter DataFrames to csv files in the directory named by SLiDE.sub_path and named for their associated dictionary key.

Arguments

dataset::String: Dataset identifier

subset::String: Internally-passed parameter indicating the type of information to save (set, parameter, or build step)

d::Dict of information to write

Returns

d::Dict: filtered dictionary

–>