Build

SLiDE.build_eem — Method

build_eem(dataset::Dataset)

If dataset.eem=true, continue the SLiDE buildstream for the Energy-Environment Module. If the dataset has already been generated and saved, the function will read and return those values.

Otherwise, it will execute the build routine via the following functions:

SLiDE.scale_sector

SLiDE.partition_seds

SLiDE.disaggregate_energy!

SLiDE.partition_co2!

SLiDE.calibrate_regional

Arguments

dataset::Dataset identifier

Returns

d::Dict of model parameters

set::Dict of Arrays describing parameter indices (years, regions, goods, sectors, etc.)

SLiDE.build_io — Method

build_io(dataset::Dataset)

If the dataset has already been generated and saved, the function will read and return those values.

Otherwise, it will read input data from the /SLIDE_DIR/data/input/ directory and execute the four steps of the SLiDE buildstream via the following functions:

SLiDE.partition_bea

SLiDE.calibrate_national

SLiDE.share_region

SLiDE.disaggregate_region

Arguments

dataset::Dataset identifier

Returns

d::Dict of model parameters

set::Dict of Arrays describing parameter indices (years, regions, goods, sectors, etc.)

SLiDE.data_saved — Method

This function returns true if parameters and sets have already been generated, and their values saved, for the given dataset.

SLiDE.datapath — Method

datapath(dataset::Dataset)

This function returns the path to the directory location specified by dataset.name/dataset.build/dataset.step. Building a dataset called dataset.name with dataset.save_build=true will produce files in the following structure. /SLIDE_DATA/data/dataset.name/ ├── eem/ | ├── parameters/ | └── sets/ ├── io/ | ├── parameters/ └───└── sets/

Arguments

dataset::String: Dataset identifier

Returns

dir::String = /path/to/dataset

SLIDE_DIR is the path to the location of the SLiDE.jl package on the user's machine.

The default dataset identifier is state_model. This dataset includes all U.S. states and summary-level sectors and goods.

SLiDE.describe — Method

describe!(set::Dict, dataset::Dataset)
describe(dataset::Dataset)

Arguments

dataset::Dataset or step::String/Symbol specifying the parameters to describe

Returns

d::Dict{Symbol,Parameter} of Parameters relevant to the specified data step. The dictionary key is consistent the value's field parameter.

SLiDE.list — Method

list!(set::Dict, dataset::Dataset)

This function adds a list of the parameters described by SLiDE.describe to set, identified by the key :step_list.

Arguments

set::Dict to update

dataset::Dataset or step::Symbol

Returns

lst::AbstractArray of parameters added to set

SLiDE.overwrite — Method

overwrite(dataset::Dataset)

This function executes dataset.overwrite=true. If a directory exists at dataset.name/dataset.build, and

Output data HAS been generated, append the date this directory was created and move it.

Output data HAS NOT yet been generated, remove the directory and start over.

SLiDE.read_build — Method

read_build(dataset::Dataset)

This function reads data from or for the specified dataset if this information has previously been generated and saved, read the saved data. If this information has NOT yet been generated, read input data using SLiDE.read_input!.

Arguments

dataset::Dataset identifier

subset::String: Internally-passed parameter indicating the type of information to save (set, parameter, or build step)

Returns

d::Dict{Symbol,DataFrame} if reading parameters or d::Dict{Any,Array} if reading sets

SLiDE.read_input! — Method

read_input!(dataset::Dataset)

Read input data for the specified dataset.build/dataset.step and set dataset.step="input to indicate further action is required.

Arguments

dataset::Dataset identifier

Returns

d::Dict of input data. If dataset.step does not require input data, return Dict().

SLiDE.read_map — Method

read_map()

Returns

d::Dict of EEM mapping datasets.

SLiDE.read_set — Method

read_set()

SLiDE.set_sector! — Method

SLiDE.write_build! — Method

write_build!(dataset::Dataset, d::Dict)

This function filters the contents of the input dictionary d to include only relevant files using SLiDE.filter_with! and writes set lists and parameter DataFrames to csv files in the directory named by SLiDE.datapath and named for their associated dictionary key.

Arguments

dataset::Dataset identifier

d::Dict of information to write

Returns

d::Dict: filtered dictionary

SLiDE.write_build — Method

write_build(path::String, k::Symbol, df::DataFrame)
write_build(path::String, k::Symbol, lst::AbstractArray)

This is a helper function for SLiDE.write_build!.

Arguments

path::String = /path/to/dataset

k: filename

df::DataFrame or lst::AbstractArray of data to write