Build

SLiDE.build_path — Method

build_path(build_step::String; kwargs...)

Arguments

build_step::String: Internally-passed parameter indicating the build step. For the four build steps, these are: partition, calibrate, share, and build.

Keyword Arguments

save: Will this data be saved? If so, where?

save = true (default): Save data in each build step (SLiDE.partition!, SLiDE.calibrate, SLiDE.share!, SLiDE.disagg!) to a directory in SLiDE/data/default/build/.

save = path/to/version: Build data will be saved in SLiDE/data/path/to/version/build/.

Returns

path::String: standardized location indicating where to save intermediate files.

SLiDE.read_build — Method

read_build(build_step::String; kwargs...)

This function reads intermediary build files if they have previously been saved by the user.

Returns

d::Dict of DataFrames

SLiDE.read_from — Method

read_from(path::String)

SLiDE.read_set — Method

read_set(path::String)

This function reads sets from the indicated paths.

Keyword

path::String: Path to location of yaml file describing which sets to store and where they are located.

Returns

d::Dict of arrays containing set information.

SLiDE.write_build — Method

write_build(build_step::String, d::Dict; kwargs...)

This function writes intermediary build files if desired by the user.

Share

SLiDE._avg_ng — Method

!!!!

SLiDE._share_d0! — Method

d0: Local supply-demand. Trade that remains within the same region.

SLiDE._share_mn0! — Method

mn0: National demand (CFS)

SLiDE._share_mrt0! — Method

mrt0: Interstate trade (CFS)

SLiDE._share_ng! — Method

ng: Sectors not included in the CFS.

SLiDE._share_xn0! — Method

xn0: National exports (CFS)

SLiDE.share_rpc! — Method

share_rpc!(d::Dict, set::Dict)

SLiDE._share_gsp! — Method

gsp: Calculated gross state product.

SLiDE._share_hw! — Method

hw(r,s): (region,sector) pairings with ALL wage shares > 1

SLiDE._share_lshr0! — Method

lshr0: Labor share of value added

SLiDE._share_wg! — Method

wg(yr,r,s): Index pairs with high wage shares (>1)

SLiDE.share_labor! — Method

share_labor!(d::Dict, set::Dict)

SLiDE.share_region! — Method

share_region!(d::Dict, set::Dict)

region: Regional share of value added

SLiDE.share_pce! — Method

share_pce!(d::Dict)

pce: Regional shares of final consumption

SLiDE.share_sgf! — Method

function share_sgf!(d::Dict)

sgf: State Government Finance data.

Note: D.C. is not included in the original data set, so assume its SGFs equal Maryland's.

SLiDE.share_utd! — Method

share_utd!(d::Dict, set::Dict)

utd: Share of total trade by region.

Disaggregate

<!– @autodocs Modules = [SLiDE] Pages = ["build/disagg.jl"] Public = false –>