Parse

Edit

SLiDE._expand_range — Method

_expand_range()

SLiDE._sort_datastream — Method

_sort_datastream(df::DataFrame)

Returns the edited DataFrame, stored in a nicely-sorted order. This is most helpful for mapping and developing. Sorting isn't necessary and we could remove this function to save some time for users.

Load

SLiDE._load_as_type — Method

_load_as_type(::Type{Any}, entry, type::DataType)

Converts an entry to the required DataType

SLiDE._load_axis — Method

_load_axis(entry::Any)

SLiDE._load_case — Method

_load_case(entry::AbstractString)

Standardizes string identifiers that indicate a case change (upper-to-lower or vice-versa) for easier editing.

SLiDE._load_path — Method

_load_path(d::Dict)

Edits directories containing a list of directories ending in a file name as one path.

Verify