Build

There are four steps required to build the cleaned input data for use in the model:

Partition

Calibrate

Share

Disaggregate

The buildstream process and all of the notation included in the SLiDE documentation here is meant to replicate the WiNDC buildstream. For more information, please reference:

Thomas F. Rutherford and Andrew Schreiber, "Tools for Open Source, Subnational CGE Modeling with an Illustrative Analysis of Carbon Leakage," J Global Econ Anal 4(2): 1-66.

SLiDE.build — Function

build(; kwargs...)

This function will execute the SLiDE buildstream and generate the parameters necessary to run the model. If the dataset has already been generated and saved, the function will read and return those values.

Otherwise, it will read BEA supply/use data from the /SLIDE_DIR/data/input/ directory.

Then, execute the four steps of the SLiDE buildstream by executing the following functions:

SLiDE.partition

SLiDE.calibrate

SLiDE.share

SLiDE.disagg

This information will be saved in the following structure:

/SLIDE_DATA/data/dataset/
├── parameters/
└── sets/

Arguments

dataset::String: Dataset identifier

Keywords

save_build::Bool = false: That decides whether to save the information at each build step. Setting save_build = true will add the directory /SLIDE_DATA/data/dataset/build to contain a subdirectory for each of the four build steps.

overwrite::Bool = false: If data exists, do not read it. Build the data from scratch.

Returns

d::Dict{Symbol,DataFrame} of model parameters

set::Dict{Any,Array} of Arrays describing parameter indices (years, regions, goods, sectors, etc.)

Partition

Divide BEA supply/use data into parameters.

SLiDE.partition — Function

partition(d::Dict, set::Dict; kwargs...)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

save_build = true

overwrite = false

See SLiDE.build for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the

Calibrate

SLiDE.calibrate — Function

calibrate(d::Dict, set::Dict; save_build=true, overwrite=false)
calibrate(year::Int, d::Dict, set::Dict)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

year::Int: year for which to perform calibration

Keywords

save_build = true

overwrite = false

See SLiDE.build for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the calibration step.

Share

Divide data into regional component. This will guide how to break the national data into regional components by state or county.

SLiDE.share — Function

share!(d::Dict, set::Dict; save_build = true, overwrite = false)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

save_build = true

overwrite = false

See SLiDE.build for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the sharing step.

Disaggregate

SLiDE.disagg — Function

function disagg!(d::Dict, set::Dict; save_build = true, overwrite = false)

Arguments

d::Dict of DataFrames containing the model data.

set::Dict of Arrays describing region, sector, final demand, etc.

Keywords

save_build = true

overwrite = false

See SLiDE.build for keyword argument descriptions.

Returns

d::Dict of DataFrames containing the model data at the disaggregation step