SEDS

Electricity

SLiDE.module_elegen! — Function

Electricity generation by source.

\[\bar{ele}_{yr,r,src,sec} = \left\{seds\left( yr,r,src,sec \right) \;\vert\; yr, \, r, \, (ff,re) \in src, \, sec = ele \right\}\]

For fossil fuels, use heatrate to convert as follows:

\[\bar{ele}_{yr,r,ff\in src,sec} \text{ [billion kWh]} = 10^3 \cdot \dfrac{\bar{ele}_{yr,r,ff\in src,sec} \text{ [trillion btu]}} {\bar{heatrate}_{yr,src} \text{ [btu/kWh]}}\]

Energy

SLiDE.module_energy! — Function

\[\bar{energy}_{yr,r,src,sec} = \left\{seds\left( yr,r,src,sec \right) \;\vert\; yr, \, r, \, e \in src, \, ed \in sec \right\}\]

SLiDE._module_energy_supply adds supply information from the electricity generation dataset output by SLiDE.module_elegen!. The following functions are used to calculate values or make adjustments to values in the energy dataset. These operations must occur in the following order:

SLiDE._module_energy_ref

SLiDE._module_energy_ind

SLiDE._module_energy_price

SLiDE._module_energy_supply — Function

\[\bar{supply}_{yr,r,src=ele} = \sum_{src} \bar{ele}_{yr,r,src}\]

SLiDE._module_energy_ref — Function

\[\bar{ref}_{yr,r,src=ele} \text{ [billion kWh]} = \bar{ref}_{yr,r,src=ele} \text{ [trillion btu]} \cdot \dfrac{\bar{ind}_{yr,r,src=ele} \text{ [billion kWh]}} {\bar{ind}_{yr,r,src=ele} \text{ [trillion btu]}}\]

SLiDE._module_energy_ind — Function

\[\bar{ind}_{yr,r,src=(ff,ele)} = \bar{ind}_{yr,r,src=(ff,ele)} - \bar{ref}_{yr,r,src=(ff,ele)}\]

SLiDE._module_energy_price — Function

\[\begin{aligned} \bar{ff}_{yr,r,sec=ele} \text{ [USD/million btu]} &= 10^3 \cdot \dfrac{\bar{ff}_{yr,r,sec=ele} \text{ [billion USD]}} {\bar{ff}_{yr,r,sec=ele} \text{ [trillion btu]}} \\ \bar{ele}_{yr,r,sec} \text{ [USD/thousand kWh]} &= 10^3 \cdot \dfrac{\bar{ele}_{yr,r,sec} \text{ [billion USD]}} {\bar{ele}_{yr,r,sec} \text{ [thousand kWh]}} \end{aligned}\]

Emissions

SLiDE.module_co2emis! — Function

Carbon dioxide emissions in million metric tons of carbon dioxide. Use SEDS data to calculate emissions by sectoral use where possible:

\[\bar{btu}_{yr,r,src,sec} = \left\{\bar{energy}\left( yr,r,src,sec \right) \;\vert\; yr, \, r, \, e \in src, \, sec \right\}\]

\[\bar{co_2}_{yr,r,src,sec} \text{ [million metric tons of CO2]} = 10^{-3} \cdot \dfrac{\bar{btu}_{yr,r,src,sec} \text{ [trillion btu]}} {{co_2/btu}_{src} \text{ [kg CO2/million btu]}}\]

Use EPA emissions data for total emissions by source.